Création de l'espace public et mise en place de la liste des animaux et accès habitat aux publics
