+++ src/Controller/HabitatController.php
@@ -243,77 +243,6 @@ final class HabitatController extends AbstractController
         );
     }
 
-    #[Route('/', name: 'index', methods: ['GET'])]
-    #[OA\Get(
-        path: "/api/habitat/",
-        summary: "Récupérer la liste des habitat visibles",
-        description: "Cette route retourne tous les habitats"
-    )]
-    #[OA\Response(
-        response: 200,
-        description: "Liste des habitat visibles",
-        content: new OA\JsonContent(
-            type: 'array',
-            items: new OA\Items(
-                properties: [
-                    new OA\Property(
-                        property: "name",
-                        type: "string",
-                        example: "Marais"
-                    ),
-                    new OA\Property(
-                        property: "description",
-                        type: "string",
-                        example: "Un habitat riche"
-                    ),
-                    new OA\Property(
-                        property: "CommentHabitat",
-                        type: "string",
-                        example: "C'est un environnement humide et froid"
-                    ),
-                    new OA\Property(
-                        property: "animal",
-                        type: "string",
-                        example: "Sama"
-                    ),
-                    new OA\Property(
-                        property: "imageUrl",
-                        type: "string",
-                        example: "https://example.com/uploads/marais.jpg"
-                    ),
-                    new OA\Property(
-                        property: "createdAt",
-                        type: "string",
-                        format: "date-time",
-                        example: "10-10-2025"
-                    )
-                ]
-            )
-        )
-    )]
-    public function index(): JsonResponse
-    {
-        $habitats = $this->repository->findAll();
-
-        $data = array_map(
-            function (Habitat $habitat) {
-                return [
-                    'name' => $habitat->getName(),
-                    'description' => $habitat->getDescription(),
-                    'commentHabitat' => $habitat->getCommentHabitat(),
-                    'animals' => $habitat->getAnimals(),
-                    'images' => $habitat->getImages(),
-                ];
-            },
-            $habitats
-        );
-
-        return new JsonResponse(
-            $data,
-            JsonResponse::HTTP_OK
-        );
-    }
-
     #[Route('/{id}', name: 'edit', methods: 'PUT')]
     #[OA\Put(
         path: "/api/habitat/{id}",
